Feat: TempMArketViewModel로 분리
