+++ app/src/main/java/dev/kichan/marketplace/ui/MainActivity.kt
@@ -9,8 +9,6 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
-import androidx.core.app.ActivityCompat
-import androidx.core.content.ContextCompat
 import androidx.core.view.WindowCompat
 import com.github.javafaker.Faker
 import com.google.firebase.messaging.FirebaseMessaging
@@ -18,7 +16,7 @@ import com.kakao.sdk.common.util.Utility
 import com.kakao.vectormap.KakaoMapSdk
 import dev.kichan.marketplace.BuildConfig
 import dev.kichan.marketplace.SingleTonViewModel
-import dev.kichan.marketplace.ui.page.TempMarketViewModel
+import dev.kichan.marketplace.viewmodel.TempMarketViewModel
 import dev.kichan.marketplace.ui.theme.MarketPlaceTheme
 import dev.kichan.marketplace.viewmodel.AuthViewModel
 import dev.kichan.marketplace.viewmodel.CouponViewModel
