+++ app/src/main/java/dev/kichan/marketplace/ui/MyApp.kt
@@ -22,7 +22,7 @@ import dev.kichan.marketplace.common.LargeCategory
 import dev.kichan.marketplace.ui.page.ApiTestPage
 import dev.kichan.marketplace.ui.page.MarketListPage
 import dev.kichan.marketplace.ui.page.ReceivedCouponsScreen
-import dev.kichan.marketplace.ui.page.TempMarketViewModel
+import dev.kichan.marketplace.viewmodel.TempMarketViewModel
 import dev.kichan.marketplace.viewmodel.AuthViewModel
 import dev.kichan.marketplace.viewmodel.CouponViewModel
 import dev.kichan.marketplace.viewmodel.MarketViewModel
